fix: security audit fixes — RNG, timing, HANDLE, key zeroing (v0.9.88)
CRIT-01: Replace Cellframe Dilithium randombytes() zero-fallback with
         qgp_randombytes() wrapper + abort on failure
HIGH-01: Add getrandom() partial-read loop on Android (match Linux)
HIGH-02: Constant-time comparison in Dilithium signature verification
HIGH-03: Change identity lock API from int to intptr_t (Windows HANDLE safety)
HIGH-10: Zero private keys with qgp_secure_memzero() before free()

+++ messenger/src/api/dna_engine_internal.h
@@ -645,7 +645,7 @@ struct dna_engine {
     /* DHT: engine uses nodus singleton (no engine-owned context) */
 
     /* Identity lock (prevents multiple engines from loading same identity) */
-    int identity_lock_fd;        /* File lock descriptor (-1 if not held) */
+    intptr_t identity_lock_fd;   /* File lock descriptor (-1 if not held) */
 
     /* Messenger backend */
     messenger_context_t *messenger;  /* Core messenger context */
