+++ shared/crypto/utils/qgp_platform_linux.c
@@ -517,7 +517,7 @@ int qgp_platform_cpu_count(void) {
  * Uses flock() for file-based locking across processes
  * ============================================================================ */
 
-int qgp_platform_acquire_identity_lock(const char *data_dir) {
+intptr_t qgp_platform_acquire_identity_lock(const char *data_dir) {
     if (!data_dir) {
         QGP_LOG_ERROR(LOG_TAG, "acquire_identity_lock: NULL data_dir");
         return -1;
@@ -547,8 +547,8 @@ int qgp_platform_acquire_identity_lock(const char *data_dir) {
     for (int attempt = 0; attempt < max_retries; attempt++) {
         if (flock(fd, LOCK_EX | LOCK_NB) == 0) {
             /* Lock acquired successfully */
-            QGP_LOG_INFO(LOG_TAG, "acquire_identity_lock: lock acquired (fd=%d, attempt=%d)",
-                         fd, attempt + 1);
+            QGP_LOG_INFO(LOG_TAG, "acquire_identity_lock: lock acquired (fd=%td, attempt=%d)",
+                         (intptr_t)fd, attempt + 1);
             return fd;
         }
 
@@ -569,7 +569,7 @@ int qgp_platform_acquire_identity_lock(const char *data_dir) {
     return -1;
 }
 
-void qgp_platform_release_identity_lock(int lock_fd) {
+void qgp_platform_release_identity_lock(intptr_t lock_fd) {
     if (lock_fd < 0) {
         return;  /* No lock to release */
     }
@@ -577,7 +577,7 @@ void qgp_platform_release_identity_lock(int lock_fd) {
     /* Release the lock and close the file descriptor */
     flock(lock_fd, LOCK_UN);
     close(lock_fd);
-    QGP_LOG_INFO(LOG_TAG, "release_identity_lock: lock released (fd=%d)", lock_fd);
+    QGP_LOG_INFO(LOG_TAG, "release_identity_lock: lock released (fd=%td)", lock_fd);
 }
 
 int qgp_platform_is_identity_locked(const char *data_dir) {
